requirements.txt
astroid==2.5.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
flake8==3.8.4
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
websocket-client~=1.2.1
Flask~=2.0.1
